change Com::send_indSstate
