+++ src/ayab/fsm.h
@@ -24,10 +24,52 @@
 #ifndef FSM_H_
 #define FSM_H_
 
-#include "com.h"
-
-enum OpState { s_wait_for_machine, s_init, s_ready, s_knit, s_test, s_error };
+enum OpState { s_wait_for_machine = 0, s_init = 1, s_ready = 2, s_knit = 3, s_test = 4, s_error = 5};
 using OpState_t = enum OpState;
+
+// As of APIv6, the only important distinction
+// is between `SUCCESS` (0) and any other value.
+// Informative error codes are provided for
+// diagnostic purposes (that is, for debugging).
+// Non-zero error codes are subject to change.
+// Such changes will be considered non-breaking.
+enum ErrorCode {
+  SUCCESS = 0x00,
+
+  // message not understood
+  EXPECTED_LONGER_MESSAGE = 0x01,
+  UNRECOGNIZED_MSGID = 0x02,
+  UNEXPECTED_MSGID = 0x03,
+  CHECKSUM_ERROR = 0x04,
+
+  // invalid arguments
+  MACHINE_TYPE_INVALID = 0x10,
+  NEEDLE_VALUE_INVALID = 0x11,
+  NULL_POINTER_ARGUMENT = 0x12,
+  ARGUMENT_INVALID = 0x13,
+  ARGUMENTS_INCOMPATIBLE = 0x13,
+
+  // device not initialized
+  NO_MACHINE_TYPE = 0x20,
+  NO_CARRIAGE = 0x21,
+  NO_DIRECTION = 0x22,
+  NO_BELTSHIFT = 0x23,
+
+  // machine in wrong FSM state
+  MACHINE_STATE_INIT = 0xE0,
+  MACHINE_STATE_READY = 0xE1,
+  MACHINE_STATE_KNIT = 0xE2,
+  MACHINE_STATE_TEST = 0xE3,
+  WRONG_MACHINE_STATE = 0xEF,
+
+  // generic error codes
+  WARNING = 0xF0, // ignorable error
+  RECOVERABLE_ERROR = 0xF1,
+  CRITICAL_ERROR = 0xF2,
+  FATAL_ERROR = 0xF3,
+  UNSPECIFIED_FAILURE = 0xFF
+};
+using Err_t = enum ErrorCode;
 
 class FsmInterface {
 public:
